Added a janky menu
Brimstone key drop potentials are now accessible through a menu with options to pick which location you are at to better determine your drop. 

Currently the biggest issue is the 20% extra chance if the monster has a slayer requirement is not implemented. This will get worked on and uploaded very soon!